+++ chronosphere/resource_monitor.go
@@ -380,6 +380,16 @@ func monitorConditionsToModel(
 				Enabled: c.ResolveValue.Enabled,
 			}
 		}
+		if c.ResolveSustainForNoData != nil {
+			secs, err := durationToSecs(c.ResolveSustainForNoData.Value)
+			if err != nil {
+				return nil, err
+			}
+			cond.ResolveSustainForNoData = &models.Configv1ResolveSustainForNoData{
+				Enabled: c.ResolveSustainForNoData.Enabled,
+				Secs:    secs,
+			}
+		}
 		bySev[c.Severity] = append(bySev[c.Severity], cond)
 	}
 
@@ -423,6 +433,12 @@ func monitorConditionsFromModel(
 				cond.ResolveValue = &intschema.MonitorSeriesConditionResolveValue{
 					Value:   c.ResolveValue.Value,
 					Enabled: c.ResolveValue.Enabled,
+				}
+			}
+			if c.ResolveSustainForNoData != nil {
+				cond.ResolveSustainForNoData = &intschema.MonitorSeriesConditionResolveSustainForNoData{
+					Enabled: c.ResolveSustainForNoData.Enabled,
+					Value:   durationFromSecs(c.ResolveSustainForNoData.Secs),
 				}
 			}
 			out = append(out, cond)